add WorkspaceDir, which sets the current directory in a workspace.
Actually, it sets the current directory in a layout, since there's no way I
know of to attach a behavior to a workspace.  This means that any terminals
(or other programs) pulled up in that workspace (with that layout) will
execute in that working directory.  Sort of handy, I think.