make combo sort of work with new doLayout.
For some reason (not entirely clear to me) this doesn't work properly just
yet with the tabbed layout.  :( But at least it'll compile.  The trouble is
that we have no way of tracking which tab ought to be visible without
adding a *lot* of infrastructure.  I'd rather have that infrastructure in
xmonad proper than reimplement all the focus-handling in combo, so for now
I'll just delay upgrading my xmonad at work...