Tabbed.hs: fixed centerText issues and some binding shadowing warnings
This patch fixes the centerText issue due to the inappropriate use of
textExtends and textWidth. Those functions need a FontStruct id to
operate, and this cannot be retrieved with queryFont (see comments in
Graphics.X11.Xlib.Font).
So we now get the FontStruct with loadQueryFont, we set the default
Xorg fonts and we calculate things for (vertical and horizontal)
centering.
It also removes some binding shadows compiler warnings