add SimpleStacking module to make Combo and Tabbed work together.
WARNING! This change will break existing Tabbed configurations.  The
problem is that there is no way within a Layout's "doLayout" to safely
modify the layout itself.  This makes LayoutHooks fragile, and more to the
point, makes SimpleStacking fragile, so we can't safely define a

tabbed' = simpleStacking . tabbed

A workaround would have been to duplicate the tabbed code, but I'd rather
leave the ugliness and get this fixed.