add CopyWindow module, to support sticky/tagged windows.
This module allows dwm-style tagging (as I understand dwm).  You can have a
given window visible in multiple workspaces.  If it's visible in two
workspaces both of which are visible, a gap will show up in one of them
(which is something that needs fixing in xmonad core).  Also defines a
kill1 which is like kill, but only removes from the current workspace a
window if it's in multiple workspaces.