NextWorkspace: Go forward or backward
Hi,

inspired by RotView, I implemented an Extension that allows the user to go
forward or backward in the list of workspaces, or to move the current
window to the next or previous workspace. Haddock included. Works here, but
hardly tested (and while tired).

Cu torrow @ HacII, if you are there.

Greetings,
Joachim