X.A.MessageFeedback fix refresh behaviour
This fixes the issue "X.L.NoBorders not working correctly together with
X.L.Groups" (#191).

Generaly when the users sends a message, they expect the layout to be
updated, not updating they layout is the exception. In the documentation
for the send method it mentioned that it worked like X.O.sendMessage,
even though it actually behaved more like X.O.sendMessageWithNoRefresh.
This PR changes this behaviour so that send actually does an update at
the end as expected.

This fixes the behaviour of X.L.NoBorders together with X.L.Groups, when
using X.L.G.Examples and X.L.G.Helpers to send messages.

I have added a new function for each old one that keeps the old
behaviour, but refreshing is the new default. The only place this module
is used in xmonad-contrib is in X.L.G.Helpers, which is the code we
actually wanted to fix.

An alternative to this is that keep the old behaviour in the existing
functions of X.A.MessageFeedback. But looking at X.O.sendMessage and
X.O.sendMessageWithNoRefresh from xmonad, this seems like the reasonable
path forward.