In XMonad.Prompt.Unicode, replace unsafePerformIO w/ ExtensibleState
Using unsafePerformIO is a sin in Haskell. In this module,
XMonad.Prompt.Unicode it was used as a dirty hack for caching the contents of
the UnicodeData file.

This patch removes the usage of unsafePerformIO within the module and instead
makes usage of XMonad.Util.ExtensibleState to save the contents of the
UnicodeData file.

Furthermore, it allows to specify a custom path to that UnicodeData
file (unicodePromptFromFile). It does not rename old identifiers though.