X.H.DynamicLog: Postpone statusBar start until after xmonad is recompiled
The sample (perhaps even recommended) usage of DynamicLog is

    main = xmonad =<< xmobar def

which invokes `spawnPipe "xmobar"` even before xmonad parses its
arguments, recompiles itself and proceeds to execute the compiled
configuration binary.

This is unnecessary, wasteful, ugly, and it occasionally triggers a bug
in xmobar, and people then waste days or even weeks chasing it as nobody
expects xmonad to start xmobar multiple times:
https://github.com/jaor/xmobar/issues/442
https://github.com/jaor/xmobar/pull/448#issuecomment-727097216
https://github.com/jaor/xmobar/pull/448#issuecomment-727161929

To fix this, we move status bar startup to the startupHook, where it
belongs, and use an IORef to pass the pipe handle to logHook.